Matthewhuang/test edited (#1)
This PR can be reviewed commit-by-commit.

1. asdf
2. asdfefefef

[LINK](https://google.com)

![image](https://user-images.githubusercontent.com/1274312/83239645-1845c200-a14d-11ea-854e-a75740bb9275.png)

Any <img> tags are bad <div and heave ho>.
< div div >
<div div >
< div div>
<div div>
<div>
< div>
<div >
< div >

![rubiks cube](https://user-images.githubusercontent.com/17112269/83487578-1c7d2280-a460-11ea-81a6-16a77f5bbfbb.gif)

hi >

> hi quote

**asdf**

_**asdf**_

- [x] one

- [ ] two

@matthewhuang97

<img width="889" alt="Screen Shot 2020-05-26 at 9 47 02 AM" src="https://user-images.githubusercontent.com/17112269/83485795-a5925a80-a45c-11ea-86e6-f90b552a6b01.png">

[test.pdf](https://github.com/matthewhuang97/test/files/4715093/test.pdf)
[test.txt](https://github.com/matthewhuang97/test/files/4715095/test.txt)

This PR enables the

This action is enabled when every node in the is either row the same rules as the single-component creation flow.

To convert a int:
- Untouched, if a
- Converted to a otherwise. If the contained, those are outside of the

_Creating blah blah blah_

(*) Note that this PR doesn't work properly without https://github.com being landed (the toolbar may act a bit funky).

Added integration tests + updated existing tests

We really don't want to be able to create a... that being said, we don't want to -- both can be addressed by enforcing the latter behavior. I'll file this + address this in a separate PR.
https://app.asana.com/0/1112356504365123056041/1178111237041329489

- Commit 1: Updates a comment for accuracy, based on my understanding. Let me know if it's inaccurate !
- Commit 2: This is the core of this PR -- it modifies the behavior for creating a from a selection + when the action is enabled.
- Commit 3: Updates fullscreen tests
- Commit 4: Adds the action to the toolbar

- Commit 5: Adds + adjusts interaction tests

https://app.asana.com/0/1156512304365056041/1177534912330360744

See the original pull request here: https://github.com/matthewhuang97/test/pull/1